requirements.txt
Django==4.2
django-csp==3.7
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
requests==2.31.0
telepot==12.7
gunicorn==20.1.0
Pillow==9.2.0
sorl-thumbnail==12.8.0
psycopg2-binary
python-dotenv==1.0.1
pytest